renamed/organized files; tested to work again; began deleteTut in store
