+++ react-app/src/components/Tuts/UploadTut/index.js
@@ -3,7 +3,7 @@
 import React, {useState} from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
-import { uploadTut } from "../../store/tuts";
+import { uploadTut } from "../../../store/tuts";
 
 
 const UploadTut = () => {
