+++ react-app/src/components/auth/LoginForm.js
@@ -2,7 +2,8 @@ import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 import { login } from '../../store/session';
-import UploadTut from '../Uploadtut';
+import UploadTut from '../Tuts/UploadTut';
+// import UploadTut from '../Uploadtut';
 // import UploadPicture from '../UploadPicture';
 
 const LoginForm = () => {
